% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{get_global_mean_norm}
\alias{get_global_mean_norm}
\title{Function to normalize raw Ct values using a global mean method.}
\usage{
get_global_mean_norm(D)
}
\arguments{
\item{D}{KxJ matrix of raw Ct values for miRNAs across samples, where row k corresponds to k-th miRNA, and column j corresponds to j-th sample. Missing values can be represented with NA.}
}
\value{
A list containing
A J-length vector of normalizers for each sample \code{u_j}
A KxJ matrix of normalized Ct values \code{D_normalized}
}
\description{
Function to normalize raw Ct values using a global mean method.
}
