% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{lr_correct_norm}
\alias{lr_correct_norm}
\title{Function to conduct correction (secondary normalization) by linear regression normalized Ct matrix.}
\usage{
lr_correct_norm(D, u_j, non_missing_thres = 5)
}
\arguments{
\item{D}{KxJ matrix of normalized Ct values for miRNAs across samples, where row k corresponds to k-th miRNA, and column j corresponds to j-th sample. Missing values can be represented with NA.}

\item{u_j}{a J length vector of normalizers across J samples obtained from global mean/fit normalization}

\item{non_missing_thres}{the minimum number of samples required for conducting linear regression correction}
}
\value{
A list containing
A KxJ matrix of corrected/secondary normalized Ct values \code{D_normalized}
A A Kx2 matrix of parameter estimate for fitted simple linear regression models for K miRNAs
}
\description{
Function to conduct correction (secondary normalization) by linear regression normalized Ct matrix.
}
